Add opt. for residual connections to 3D UNet
