+++ neurofire/models/unet/unet_3d.py
@@ -1,5 +1,8 @@
 import torch.nn as nn
-from .base import UNetSkeleton, Xcoder
+from inferno.extensions.layers import ConvELU3D
+from torch import nn as nn
+
+from .base import UNetSkeleton, Xcoder, XcoderResidual
 from inferno.extensions.layers.convolutional import ConvELU3D, Conv3D, BNReLUConv3D
 from inferno.extensions.layers.sampling import AnisotropicPool, AnisotropicUpsample
 
@@ -47,6 +50,53 @@ class Base(Xcoder):
                                    pre_output=None)
 
 
+
+
+
+class EncoderResidual(XcoderResidual):
+    def __init__(self, in_channels, out_channels, kernel_size, scale_factor=2, conv_type=ConvELU3D):
+        assert isinstance(scale_factor, (int, list, tuple))
+        if isinstance(scale_factor, (list, tuple)):
+            assert len(scale_factor) == 3
+            # we need to make sure that the scale factor conforms with the single value
+            # that AnisotropicPool expects
+            assert scale_factor[0] == 1
+            assert scale_factor[1] == scale_factor[2]
+            sampler = AnisotropicPool(downscale_factor=scale_factor[1])
+        else:
+            sampler = nn.MaxPool3d(kernel_size=1 + scale_factor,
+                                   stride=scale_factor,
+                                   padding=1)
+        super(EncoderResidual, self).__init__(in_channels, out_channels, kernel_size,
+                                              conv_type=conv_type,
+                                              pre_output=sampler)
+
+
+class DecoderResidual(XcoderResidual):
+    def __init__(self, in_channels, out_channels, kernel_size, scale_factor=2, conv_type=ConvELU3D):
+        assert isinstance(scale_factor, (int, list, tuple))
+        if isinstance(scale_factor, (list, tuple)):
+            assert len(scale_factor) == 3
+            # we need to make sure that the scale factor conforms with the single value
+            # that AnisotropicPool expects
+            assert scale_factor[0] == 1
+            assert scale_factor[1] == scale_factor[2]
+            sampler = AnisotropicUpsample(scale_factor=scale_factor[1])
+        else:
+            sampler = nn.Upsample(scale_factor=scale_factor)
+        super(DecoderResidual, self).__init__(in_channels, out_channels, kernel_size,
+                                              conv_type=conv_type,
+                                              pre_output=sampler)
+
+
+
+class BaseResidual(XcoderResidual):
+    def __init__(self, in_channels, out_channels, kernel_size, conv_type=ConvELU3D):
+        super(BaseResidual, self).__init__(in_channels, out_channels, kernel_size,
+                                           conv_type=conv_type,
+                                           pre_output=None)
+
+
 class Output(Conv3D):
     def __init__(self, in_channels, out_channels, kernel_size):
         super(Output, self).__init__(in_channels, out_channels, kernel_size)
@@ -67,7 +117,8 @@ class UNet3D(UNetSkeleton):
                  fmap_growth,
                  scale_factor=2,
                  final_activation='auto',
-                 conv_type_key='vanilla'):
+                 conv_type_key='vanilla',
+                 add_residual_connections=False):
         """
         Parameter:
         ----------
@@ -99,30 +150,33 @@ class UNet3D(UNetSkeleton):
         self.out_channels = out_channels
 
         conv_type = CONV_TYPES[conv_type_key]
+        decoder_type = DecoderResidual if add_residual_connections else Decoder
+        encoder_type = EncoderResidual if add_residual_connections else Encoder
+        base_type = BaseResidual if add_residual_connections else Base
 
         # Build encoders with proper number of feature maps
         f0e = initial_num_fmaps
         f1e = initial_num_fmaps * fmap_growth
         f2e = initial_num_fmaps * fmap_growth**2
         encoders = [
-            Encoder(in_channels, f0e, 3, self.scale_factor[0], conv_type=conv_type),
-            Encoder(f0e, f1e, 3, self.scale_factor[1], conv_type=conv_type),
-            Encoder(f1e, f2e, 3, self.scale_factor[2], conv_type=conv_type)
+            encoder_type(in_channels, f0e, 3, self.scale_factor[0], conv_type=conv_type),
+            encoder_type(f0e, f1e, 3, self.scale_factor[1], conv_type=conv_type),
+            encoder_type(f1e, f2e, 3, self.scale_factor[2], conv_type=conv_type)
         ]
 
         # Build base
         # number of base output feature maps
         f0b = initial_num_fmaps * fmap_growth**3
-        base = Base(f2e, f0b, 3, conv_type=conv_type)
+        base = base_type(f2e, f0b, 3, conv_type=conv_type)
 
         # Build decoders (same number of feature maps as MALA)
         f2d = initial_num_fmaps * fmap_growth**2
         f1d = initial_num_fmaps * fmap_growth
         f0d = initial_num_fmaps
         decoders = [
-            Decoder(f0b + f2e, f2d, 3, self.scale_factor[2], conv_type=conv_type),
-            Decoder(f2d + f1e, f1d, 3, self.scale_factor[1], conv_type=conv_type),
-            Decoder(f1d + f0e, f0d, 3, self.scale_factor[0], conv_type=conv_type)
+            decoder_type(f0b + f2e, f2d, 3, self.scale_factor[2], conv_type=conv_type),
+            decoder_type(f2d + f1e, f1d, 3, self.scale_factor[1], conv_type=conv_type),
+            decoder_type(f1d + f0e, f0d, 3, self.scale_factor[0], conv_type=conv_type)
         ]
 
         # Build output
